remove categories from account module
