+++ app/models/category.rb
@@ -1,0 +1,5 @@
+class Category < ApplicationRecord
+  belongs_to :user
+
+  validates :name, presence: true, uniqueness: true
+end
