--- app/views/categories/_category.json.jbuilder
@@ -1,2 +1,0 @@
-json.extract! category, :id, :name, :created_at, :updated_at
-json.url category_url(category, format: :json)
