--- app/views/categories/show.html.erb
@@ -1,19 +1,0 @@
-<p class="title">
-  <strong>Name:</strong>
-  <%= @category.name %>
-</p>
-
-<table class="table table-hover">
-  <thead>
-    <tr>
-      <th>Name</th>
-    </tr>
-  <tbody>
-      <tr>
-        <td><%= @category.name %></td>
-      </tr>
-    <tbody>
-</table>
-
-<%= link_to 'Edit', edit_category_path(@category), :class => "btn btn-warning" %>
-<%= link_to 'Back', categories_path, :class => "btn btn-primary" %>
