--- app/views/categories/show.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! 'categories/category', category: @category
