+++ spec/requests/categories_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe 'Category', type: :request do
+RSpec.describe Category, type: :request do
   let(:user) { create(:user) }
   let!(:category) { create(:category, user: user) }
